SQUASH ME: add impl segment to package; bump TC to 1.16.0"
