+++ test-utils/testcontainers/src/main/resources/META-INF/services/org.testcontainers.containers.JdbcDatabaseContainerProvider
@@ -1,5 +1,5 @@
-org.camunda.test.utils.testcontainers.CamundaPostgreSQLContainerProvider
-org.camunda.test.utils.testcontainers.CamundaMariaDBContainerProvider
-org.camunda.test.utils.testcontainers.CamundaMySqlContainerProvider
-org.camunda.test.utils.testcontainers.CamundaCockroachDBContainerProvider
-org.camunda.test.utils.testcontainers.CamundaMSSQLContainerProvider
+org.camunda.impl.test.utils.testcontainers.CamundaPostgreSQLContainerProvider
+org.camunda.impl.test.utils.testcontainers.CamundaMariaDBContainerProvider
+org.camunda.impl.test.utils.testcontainers.CamundaMySqlContainerProvider
+org.camunda.impl.test.utils.testcontainers.CamundaCockroachDBContainerProvider
+org.camunda.impl.test.utils.testcontainers.CamundaMSSQLContainerProvider
